Phase 2 Core Components: Variable packets, multi-byte checksums, rich progress
🚀 Major Phase 2 Achievements:

Variable Packet Support:
- Packet data loader supporting 1-1024 byte packets (vs fixed 7-byte)
- Multi-byte checksum support (1-8 bytes: CRC16, CRC32, CRC64)
- Multiple input formats (CSV, JSON, HEX) for real-world data
- Backwards compatible GMRS dataset for testing

Algorithm Framework Upgrade:
- All 29 algorithms upgraded for uint64_t multi-byte operation
- Type-safe interfaces with proper checksum size handling
- Enhanced CRC, Fletcher, and polynomial implementations
- Bit-width aware rotation and bit manipulation functions

Professional Progress Tracking:
- Rich progress bars with colored terminal output
- ETA calculations with accurate time estimates
- Performance metrics (tests/sec, elapsed time, solutions found)
- Professional summaries with helpful suggestions when no solutions found

Field Processing Infrastructure:
- Field combination generator for systematic packet exploration
- Efficient permutation handling with performance safeguards
- Validation and error checking for robust operation

Next: Core search engine with early exit logic and threading support

+++ src/algorithms/intermediate_ops.h
@@ -4,18 +4,18 @@
 #include "../../include/cads_types.h"
 
 // Intermediate algorithm implementations (12 operations)
-uint8_t intermediate_not(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_lshift(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_rshift(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_mul(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_div(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_mod(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_negate(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_const_add(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_const_xor(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_const_sub(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_ones_complement(uint8_t a, uint8_t b, uint8_t constant);
-uint8_t intermediate_twos_complement(uint8_t a, uint8_t b, uint8_t constant);
+uint64_t intermediate_not(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_lshift(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_rshift(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_mul(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_div(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_mod(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_negate(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_const_add(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_const_xor(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_const_sub(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_ones_complement(uint64_t a, uint64_t b, uint64_t constant);
+uint64_t intermediate_twos_complement(uint64_t a, uint64_t b, uint64_t constant);
 
 // Registry function for intermediate algorithms
 const algorithm_info_t* get_intermediate_algorithms(int* count);
